+++ anyblok_wms_base/core/goods/type.py
@@ -10,7 +10,12 @@
 from anyblok import Declarations
 from anyblok.column import String
 from anyblok.column import Integer
+from anyblok.relationship import Many2One
 from anyblok_postgres.column import Jsonb
+from anyblok_wms_base.utils import dict_merge
+
+_missing = object()
+"""A marker to use as default value in get-like functions/methods."""
 
 
 register = Declarations.register
@@ -57,21 +62,111 @@ class Type:
     using constants from an autodocumented module)
     """
 
+    properties = Jsonb(label="Properties")
+    """Goods Types also have flexible properties.
+
+    These are usually read from the Goods themselves (where they act as
+    default values if not defined on the Goods), and are useful with
+    generic Types, i.e., those that have children. Operations that handle
+    Properties can do interesting things by using properties that actually
+    come from Type information.
+    """
+
+    parent = Many2One(model='Model.Wms.Goods.Type')
+    """This field expresses the hierarchy of Goods Types."""
+
     def __str__(self):
         return "(id={self.id}, code={self.code!r})".format(self=self)
 
     def __repr__(self):
         return "Wms.Goods.Type" + str(self)
 
+    # TODO PERF cache ?
     def get_behaviour(self, name, default=None):
         """Get the value of the behaviour with given name.
 
         This method is the preferred way to access a given behaviour.
-        It performs all the needed resolutions and defaultings.
-        In particular, it takes care of the case where :attr:`behaviours` is
+        It resolves the wished behaviour by looking it up within the
+        :attr:`behaviours` :class:`dict`, and recursively on its parent.
+
+        It also takes care of corner cases, such as when :attr:`behaviours` is
         ``None`` as a whole.
         """
         behaviours = self.behaviours
+        parent = self.parent
+
+        if parent is None:
+            parent_beh = _missing
+        else:
+            parent_beh = self.parent.get_behaviour(name, default=_missing)
+
         if behaviours is None:
-            return default
-        return behaviours.get(name, default)
+            beh = _missing
+        else:
+            beh = behaviours.get(name, _missing)
+
+        if beh is _missing:
+            if parent_beh is _missing:
+                return default
+            return parent_beh
+        if parent_beh is _missing:
+            return beh
+        return dict_merge(beh, parent_beh)
+
+    def is_sub_type(self, gt):
+        """True if ``self``  is a sub type of ``gt``, inclusively.
+
+        TODO PERF the current implementation recurses over ancestors.
+        A subsequent implementation could add caching and/or recursive SQL
+        queries.
+        """
+        if self == gt:
+            return True
+        parent = self.parent
+        if parent is None:
+            return False
+        return parent.is_sub_type(gt)
+
+    def get_property(self, k, default=None):
+        """Read a property value recursively.
+
+        If the current Type does not have the wished property key, but has a
+        parent, then the lookup continues on the parent.
+        """
+        props = self.properties
+        val = _missing if props is None else props.get(k, _missing)
+        if val is _missing:
+            parent = self.parent
+            if parent is None:
+                return default
+            return parent.get_property(k, default=default)
+        return val
+
+    def has_property_values(self, mapping):
+        return all(self.get_property(k, default=_missing) == v
+                   for k, v in mapping.items())
+
+    def has_property(self, name):
+        if self.properties is not None and name in self.properties:
+            return True
+        parent = self.parent
+        if parent is not None:
+            return parent.has_property(name)
+        return False
+
+    def has_properties(self, wanted_props):
+        if not wanted_props:
+            return True
+
+        properties = self.properties
+        if properties is None:
+            missing = wanted_props
+        else:
+            missing = (p for p in wanted_props if p not in properties)
+
+        parent = self.parent
+        if parent is None:
+            for x in missing:  # could be a generator, a list etc.
+                return False
+            return True
+        return parent.has_properties(missing)